backend/requirements.txt
strands-agents[bidi]>=1.28.0
strands-agents-tools>=0.1.0
mcp>=1.0.0
boto3>=1.35.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
hypothesis
